pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "investment-research-rag"
version = "0.1.0"
description = "Retrieval-augmented generation for institutional investment research — SEC filings, earnings transcripts, and equity research reports"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "bdschi1"},
]
keywords = [
    "rag", "retrieval-augmented-generation", "finance", "investment-research",
    "sec-filings", "earnings-transcripts", "vector-search", "embeddings",
    "llm-evaluation", "financial-reasoning",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "tiktoken>=0.7.0",
    "pdfplumber>=0.10.0",
    "python-docx>=1.0.0",
    "pandas>=2.0.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
faiss = ["faiss-cpu>=1.7.4"]
qdrant = ["qdrant-client>=1.7.0"]
openai = ["openai>=1.50.0"]
anthropic = ["anthropic>=0.40.0"]
ollama = ["ollama>=0.4.0"]
huggingface = ["sentence-transformers>=2.2.0", "torch>=2.0"]
reranker = ["sentence-transformers>=2.2.0"]
sec = ["pymupdf>=1.24.0"]
excel = ["openpyxl>=3.1.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "fpdf2>=2.7.0",
]
local = [
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=2.2.0",
    "ollama>=0.4.0",
]
all = [
    "faiss-cpu>=1.7.4",
    "qdrant-client>=1.7.0",
    "openai>=1.50.0",
    "anthropic>=0.40.0",
    "ollama>=0.4.0",
    "sentence-transformers>=2.2.0",
    "torch>=2.0",
    "pymupdf>=1.24.0",
    "openpyxl>=3.1.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "fpdf2>=2.7.0",
]

[project.scripts]
rag = "cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/rag", "cli"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "."]
